requirements_backup_20250629_125425.txt
accelerate==1.6.0
acres==0.3.0
aioconsole==0.8.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioresponses==0.7.8
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
appdirs==1.4.4
art==6.5
attrs==25.3.0
Authlib==1.5.2
autoflake==2.3.1
azure-ai-contentsafety==1.0.0
azure-core==1.34.0
azure-identity==1.22.0
azure-storage-blob==12.25.1
backoff==2.2.1
bandit==1.8.6
base2048==0.1.3
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
blis==1.3.0
boolean.py==5.0
boto3==1.38.13
botocore==1.38.13
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.21.0
cohere==5.15.0
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
confusable-homoglyphs==3.3.1
confusables==1.2.0
contourpy==1.3.2
coverage==7.9.1
cryptography==44.0.3
cssselect2==0.8.0
cycler==0.12.1
cyclonedx-python-lib==9.1.0
cymem==2.0.11
datasets==3.6.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
docling==2.30.0
docling-core==2.28.0
docling-ibm-models==3.4.2
docling-parse==4.0.1
docstring_parser==0.16
dparse==0.6.4
duckdb==1.2.2
duckdb_engine==0.17.0
easyocr==1.7.2
ecdsa==0.19.1
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
et_xmlfile==2.0.0
etelemetry==0.3.1
execnet==2.1.1
factory_boy==3.3.3
Faker==37.3.0
fastapi==0.115.12
fastavro==1.10.0
filelock==3.16.1
filetype==1.2.0
fitz==0.0.1.dev2
flake8==7.3.0
fonttools==4.58.0
fpdf2==2.8.3
freezegun==1.5.2
frozenlist==1.6.0
fsspec==2025.3.0
garak==0.9.0.4
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.92.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.14.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
isodate==0.6.1
isort==6.0.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.0
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
latex2mathml==3.77.0
lazy_loader==0.4
license-expression==30.4.3
limits==5.2.0
looseversion==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.3
mccabe==0.7.0
mcp==1.9.4
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.1
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
mypy_extensions==1.1.0
narwhals==1.38.2
neo4j==5.28.1
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
nltk==3.9.1
numdifftools==0.9.41
numpy==2.2.5
ollama==0.5.3
openai==1.78.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packageurl-python==0.17.1
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathlib==1.0.1
pathspec==0.12.1
patsy==1.0.1
pbr==6.1.1
peft==0.15.2
pillow==11.2.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
plotly==6.0.1
pluggy==1.5.0
preshed==3.0.9
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
prov==2.0.1
psutil==6.1.1
puremagic==1.28
py-serializable==2.0.0
pyarrow==20.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycodestyle==2.14.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.4
pydyf==0.11.0
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
PyMuPDF==1.25.5
pyodbc==5.2.0
pyparsing==3.2.3
pypdf==5.5.0
pypdfium2==4.30.1
pyphen==0.17.2
pyrit==0.9.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-env==1.1.5
pytest-httpx==0.35.0
pytest-mock==3.14.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
replicate==1.0.6
requests==2.32.4
responses==0.25.7
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
rtree==1.4.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.12.0
safetensors==0.5.3
safety==3.6.0
safety-schemas==0.0.14
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
segno==1.6.6
semchunk==2.2.2
semopy==2.3.11
sentence-transformers==4.1.0
setuptools==79.0.0
shapely==2.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
slowapi==0.1.9
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
sse-starlette==2.3.6
starlette==0.46.2
statsmodels==0.14.4
stevedore==5.4.1
streamlit==1.45.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tinyhtml5==2.0.0
tinytag==2.1.1
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
transformers==4.51.3
treelib==1.7.1
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uv==0.6.16
uvicorn==0.34.3
uvloop==0.21.0
wasabi==1.1.3
weasel==0.4.1
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
XlsxWriter==3.2.3
xxhash==3.5.0
yamllint==1.37.1
yarl==1.20.0
zopfli==0.2.3.post1
